Remove and restructure utility methods from IAppDataService and move helpers in Core project to Utility namespace
