+++ WoWsShipBuilder.UI/Converters/ClassToPathConverter.cs
@@ -3,7 +3,7 @@ using System.Globalization;
 using Avalonia.Data;
 using Avalonia.Data.Converters;
 using Avalonia.Media;
-using WoWsShipBuilder.Core.ConverterHelpers;
+using WoWsShipBuilder.Core.Utility;
 using WoWsShipBuilder.DataStructures;
 using Brushes = Avalonia.Media.Brushes;
