+++ WoWsShipBuilder.Web/Components/ConsumableSelector.razor
@@ -2,8 +2,8 @@
 @using WoWsShipBuilder.DataStructures
 @using WoWsShipBuilder.ViewModels.ShipVm
 @using WoWsShipBuilder.Core.DataContainers
+@using WoWsShipBuilder.Core.Utility
 @using WoWsShipBuilder.DataElements.DataElements
-@using WoWsShipBuilder.Core.ConverterHelpers
 @inherits ReactiveComponentBase<ConsumableViewModel>
 @inject ILocalizer Localizer
