+++ WoWsShipBuilder.Web/Components/SharedFragments.razor
@@ -1,5 +1,5 @@
 ﻿@using WoWsShipBuilder.DataElements.DataElements
-@using WoWsShipBuilder.Core.ConverterHelpers
+@using WoWsShipBuilder.Core.Utility
 
 @code {
