+++ WoWsShipBuilder.Web/Components/ShipComparisonSectionsRenderFragments/AerialTorpedoesFragment.razor
@@ -1,6 +1,6 @@
 ﻿@using WoWsShipBuilder.ViewModels.Helper
 @using WoWsShipBuilder.ViewModels.Other
-@using WoWsShipBuilder.Core.ConverterHelpers
+@using WoWsShipBuilder.Core.Utility
 @using WoWsShipBuilder.DataStructures
 @code {
