+++ WoWsShipBuilder.Web/Components/ShipComparisonSectionsRenderFragments/TorpedoesFragment.razor
@@ -1,7 +1,7 @@
 ﻿@using WoWsShipBuilder.ViewModels.Helper
 @using WoWsShipBuilder.ViewModels.Other
 @using System.Globalization
-@using WoWsShipBuilder.Core.ConverterHelpers
+@using WoWsShipBuilder.Core.Utility
 @using WoWsShipBuilder.DataStructures
 @code {
