+++ WoWsShipBuilder.Web/Components/ShipStatsComponent.razor
@@ -1,5 +1,4 @@
-﻿@using WoWsShipBuilder.Core.ConverterHelpers
-@using WoWsShipBuilder.Core.DataContainers
+﻿@using WoWsShipBuilder.Core.DataContainers
 @using WoWsShipBuilder.Core.Localization
 @using WoWsShipBuilder.DataElements.DataElements
 @using WoWsShipBuilder.DataStructures
@@ -8,6 +7,7 @@
 @using Microsoft.Extensions.Logging.Console
 @using System.Text
 @using WoWsShipBuilder.Core.Settings
+@using WoWsShipBuilder.Core.Utility
 @using WoWsShipBuilder.DataStructures.Ship.Components
 @using WoWsShipBuilder.Web.Data
 @using WoWsShipBuilder.Web.Dialogs
