+++ WoWsShipBuilder.Web/Dialogs/ShipBuildContainerSelectionDialog.razor
@@ -2,12 +2,10 @@
 @using System.Collections.ObjectModel
 @using WoWsShipBuilder.Core.Data
 @using WoWsShipBuilder.Core.DataProvider
-@using WoWsShipBuilder.Core.Services
 @using WoWsShipBuilder.ViewModels.Other
 
 @implements IDisposable
 @inject ILocalizer Localizer
-@inject IAppDataService WebAppDataService
 
 <MudDialog ContentStyle="max-height: calc(90vh - 100px); overflow-y: scroll; padding: 8px 16px;">
     <DialogContent>
@@ -95,7 +93,7 @@
         if (e.NewItems![0] is string shipIndex)
         {
             var shipSummary = AppData.ShipSummaryList.First(x => x.Index.Equals(shipIndex));
-            var ship = WebAppDataService.GetShipFromSummary(shipSummary);
+            var ship = AppData.FindShipFromSummary(shipSummary);
 
             var shipContainer = ShipBuildContainer.CreateNew(ship, null, null);
             ShipList!.Add(shipContainer);
