+++ WoWsShipBuilder.Web/Pages/TestPage2.razor
@@ -1,11 +1,8 @@
 ﻿@page "/test2"
 @using WoWsShipBuilder.Web.Dialogs
 @using WoWsShipBuilder.Core.DataProvider
-@using WoWsShipBuilder.Core.Services
 @using WoWsShipBuilder.DataStructures
 @using WoWsShipBuilder.Core.Data
-
-@inject IAppDataService AppDataService
 
 <MudRadioGroup @bind-SelectedOption="selectedTier">
     @for (var i = 1; i <= 11; i++)
@@ -61,7 +58,7 @@
         var shipList = AppData.ShipSummaryList
             .Where(s => s.Tier == selectedTier && s.ShipClass == selectedClass)
             .Take(25)
-            .Select(summary => AppDataService.GetShipFromSummary(summary));
+            .Select(AppData.FindShipFromSummary);
         ships = shipList.Select(s => ShipBuildContainer.CreateNew(s, null, null)).ToList();
         StateHasChanged();
     }
